pyproject.toml
[project]
name = "flask-health-template"
version = "0.1.0"
description = "Flask Healthcare app boilerplate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-migrate>=4.0.0",
    "psycopg[binary]>=3.2.0",
    "psycopg2-binary>=2.9.0",
    "opensearch-py>=2.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "cryptography>=42.0.0",
    "gunicorn>=22.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]
